# LibGfx/CCITT: Make decoding of wide ccitt3-2d and ccitt-4 images faster